wrap the console's input method with the to_thread function
* moved the bot#to_thread function to console.py
